move the common comparator into JniUtils.h. sparkle-plugin#6.
